\name{BetaBit-package}
\alias{BetaBit-package}
\alias{BetaBit}
\docType{package}
\title{
Zbiory danych do opowiadan o Beta i Bit
}
\description{
Zbiory danych do opowiadan o Beta i Bit
}
\details{
\tabular{ll}{
Package: \tab BetaBit\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-29\cr
License: \tab GPL-3\cr
}
}
\author{
Przmyslaw Biecek

Maintainer: Przmyslaw Biecek <Przmyslaw.Biecek@gmail.com>
}
\keyword{ package }
