\name{greatRace}
\alias{greatRace}
\docType{data}
\title{
Data set from the story ,,Great race''
}
\description{
Data set from the story ,,Great race''
}
\usage{data(oceny)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{bike}}{time in seconds}
    \item{\code{jogging}}{time in seconds}
    \item{\code{metro}}{time in seconds}
  }
}
\details{
Data set from the story ,,Great race''
}
\examples{
data(wielkiWyscig)
}
\keyword{datasets}
