\name{wielkiWyscig}
\alias{wielkiWyscig}
\docType{data}
\title{
Zbiór danych do opowiadania ,,Wielki wyscig''
}
\description{
Zbiór danych do opowiadania ,,Wielki wyscig''
}
\usage{data(oceny)}
\format{
  A data frame with 10 observations on the following 3 variables.
  \describe{
    \item{\code{rower}}{time in seconds}
    \item{\code{bieg}}{time in seconds}
    \item{\code{metro}}{time in seconds}
  }
}
\details{
Zbiór danych do opowiadania ,,Wielki wyscig''
}
\examples{
data(wielkiWyscig)
}
\keyword{datasets}
