\name{oceny}
\alias{oceny}
\docType{data}
\title{
Zbior danych do opowiadania ,,Sposob na polonistke''
}
\description{
Zbior danych do opowiadania ,,Sposob na polonistke''
}
\usage{data(oceny)}
\format{
  A data frame with 120 observations on the following 6 variables.
  \describe{
    \item{\code{dzien}}{a character vector}
    \item{\code{godzina}}{a character vector}
    \item{\code{klasa}}{a factor with levels \code{A} \code{B}}
    \item{\code{imie}}{a character vector}
    \item{\code{nazwisko}}{a character vector}
    \item{\code{ocena}}{a numeric vector}
  }
}
\details{
Zbiór danych do opowiadania ,,Sposób na polonistke''
}
\source{
Zbiór danych do opowiadania ,,Sposób na polonistke''
}
\examples{
data(oceny)
}
\keyword{datasets}
